% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padicmetric.r
\name{padicDimred}
\alias{padicDimred}
\title{Applies dimensionality reduction methods to p-adic transformed data from scRNAseq.}
\usage{
padicDimred(data, log2Transform = FALSE, reg_eval = TRUE, method = "none",
  ndims = 2, tsne.iter = 1000, ncores = 1)
}
\arguments{
\item{data}{(numeric) data matrix with features as rows and items as columns.}

\item{log2Transform}{(logica) Apply log2 transformation to scRNAseq data.}

\item{method}{(character) Method or list of methods for application of dimensionality reduction.}

\item{tsne.iter}{(numeric) Number of iterations for tsne or Rtsne.}

\item{ncores}{(Integer) Number of cores. Default is 1.}

\item{reg.eval}{(logical) Compute R^2 to evaluate fitting of logarithmic approximation. Default is TRUE.}

\item{ndim}{(numeric) Number of components to extract from dimensionaity reduction methods.}
}
\value{
....
}
\description{
Applies dimensionality reduction methods to p-adic transformed data from scRNAseq.
}
\examples{

test ...

}
\author{
Ivan Imaz \email{ii236@cam.ac.uk}
}
